dev-to-prod-gantt-4/gantt/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
plotly==5.3.1
openpyxl==3.0.9
pytest==6.2.5
dataclasses==0.6